requirements.txt
# Core dependencies
joblib>=1.1.0
numba==0.56.4
numpy==1.23.5
scipy
librosa==0.9.1
llvmlite==0.39.0
scikit-learn
tqdm>=4.63.1
PyYAML>=6.0
python-dotenv>=1.0.0

# Audio/Video processing
pydub>=0.25.1
soundfile>=0.12.1
ffmpeg-python>=0.2.0
pyworld==0.3.2
av
resampy>=0.4.2

# Deep Learning / AI
torch==2.5.1
torchaudio==2.5.1
torchcrepe==0.0.20
torchfcpe
fairseq==0.12.2
faiss-cpu==1.7.1.post3

# RVC and TTS
piper-tts==1.3.0
# faster-whisper>=0.10.0  # Removed: now using integrated TTS+caption generation

# Image processing
Pillow>=9.1.1

# Web and API
requests>=2.25.0
beautifulsoup4>=4.9.0
click>=8.0.0
pydantic>=2.0.0
rich>=13.0.0

# Platform specific
onnxruntime; sys_platform == 'darwin'
onnxruntime-gpu; sys_platform != 'darwin'
